Fixing tests and pylint, mypy issues
