+++ src/fidesops/ops/schemas/saas/strategy_configuration.py
@@ -113,20 +113,7 @@ class QueryParamAuthenticationConfiguration(StrategyConfiguration):
     value: str
 
 
-class OAuth2AuthenticationCodeConfiguration(StrategyConfiguration):
-    """
-    OAuth2 endpoints for authentication, token retrieval, and token refresh.
-    Includes an optional expires_in parameter (in seconds) for OAuth2 integrations that
-    do not specify a TTL for the access tokens.
-    """
-
-    expires_in: Optional[int]
-    authorization_request: SaaSRequest
-    token_request: SaaSRequest
-    refresh_request: Optional[SaaSRequest]
-
-
-class OAuth2ClientCredentialsConfiguration(StrategyConfiguration):
+class OAuth2BaseConfiguration(StrategyConfiguration):
     """
     OAuth2 endpoints for token retrieval, and token refresh.
     Includes an optional expires_in parameter (in seconds) for OAuth2 integrations that
@@ -136,3 +123,12 @@ class OAuth2ClientCredentialsConfiguration(StrategyConfiguration):
     expires_in: Optional[int]
     token_request: SaaSRequest
     refresh_request: Optional[SaaSRequest]
+
+
+class OAuth2AuthenticationCodeConfiguration(OAuth2BaseConfiguration):
+    """
+    The standard OAuth2 configuration but with an additional property to configure
+    the authorization request for the Authorization Code flow.
+    """
+
+    authorization_request: SaaSRequest
