+++ src/fidesops/ops/service/authentication/authentication_strategy_oauth2_client_credentials.py
@@ -5,7 +5,7 @@ from requests import PreparedRequest
 
 from fidesops.ops.models.connectionconfig import ConnectionConfig
 from fidesops.ops.schemas.saas.strategy_configuration import (
-    OAuth2ClientCredentialsConfiguration,
+    OAuth2BaseConfiguration,
     StrategyConfiguration,
 )
 from fidesops.ops.service.authentication.authentication_strategy_oauth2_base import (
@@ -23,11 +23,6 @@ class OAuth2ClientCredentialsAuthenticationStrategy(OAuth2AuthenticationStrategy
 
     strategy_name = "oauth2_client_credentials"
 
-    def __init__(self, configuration: OAuth2ClientCredentialsConfiguration):
-        self.expires_in = configuration.expires_in
-        self.token_request = configuration.token_request
-        self.refresh_request = configuration.refresh_request
-
     def add_authentication(
         self, request: PreparedRequest, connection_config: ConnectionConfig
     ) -> PreparedRequest:
@@ -39,17 +34,17 @@ class OAuth2ClientCredentialsAuthenticationStrategy(OAuth2AuthenticationStrategy
         access_token = connection_config.secrets.get("access_token")  # type: ignore
         if not access_token:
             access_token = self.get_access_token(connection_config, None)
-        elif self.refresh_request:
+        else:
             access_token = self._refresh_token(connection_config)
 
         # add access_token to request
         request.headers["Authorization"] = "Bearer " + access_token
         return request
 
-    @staticmethod
-    def _required_secrets() -> List[str]:
+    @property
+    def _required_secrets(self) -> List[str]:
         return ["client_id", "client_secret"]
 
     @staticmethod
     def get_configuration_model() -> StrategyConfiguration:
-        return OAuth2ClientCredentialsConfiguration  # type: ignore
+        return OAuth2BaseConfiguration  # type: ignore
