+++ tests/ops/fixtures/saas_example_fixtures.py
@@ -13,7 +13,6 @@ from fidesops.ops.models.connectionconfig import (
 from fidesops.ops.models.datasetconfig import DatasetConfig
 from fidesops.ops.schemas.saas.strategy_configuration import (
     OAuth2AuthenticationCodeConfiguration,
-    OAuth2ClientCredentialsConfiguration,
 )
 from fidesops.ops.util.saas_util import load_config
 from tests.ops.fixtures.application_fixtures import load_dataset
@@ -128,7 +127,7 @@ def saas_example_connection_config_with_invalid_saas_config(
 
 
 @pytest.fixture(scope="function")
-def oauth2_client_credentials_configuration() -> OAuth2ClientCredentialsConfiguration:
+def oauth2_authorization_code_configuration() -> OAuth2AuthenticationCodeConfiguration:
     return {
         "authorization_request": {
             "method": "GET",
@@ -179,6 +178,53 @@ def oauth2_client_credentials_configuration() -> OAuth2ClientCredentialsConfigur
             ],
         },
     }
+
+
+@pytest.fixture(scope="function")
+def oauth2_authorization_code_connection_config(
+    db: Session, oauth2_authorization_code_configuration
+) -> Generator:
+    secrets = {
+        "domain": "localhost",
+        "client_id": "client",
+        "client_secret": "secret",
+        "redirect_uri": "https://localhost/callback",
+        "access_token": "access",
+        "refresh_token": "refresh",
+    }
+    saas_config = {
+        "fides_key": "oauth2_authorization_code_connector",
+        "name": "OAuth2 Auth Code Connector",
+        "type": "custom",
+        "description": "Generic OAuth2 connector for testing",
+        "version": "0.0.1",
+        "connector_params": [{"name": item} for item in secrets.keys()],
+        "client_config": {
+            "protocol": "https",
+            "host": secrets["domain"],
+            "authentication": {
+                "strategy": "oauth2_authorization_code",
+                "configuration": oauth2_authorization_code_configuration,
+            },
+        },
+        "endpoints": [],
+        "test_request": {"method": "GET", "path": "/test"},
+    }
+
+    fides_key = saas_config["fides_key"]
+    connection_config = ConnectionConfig.create(
+        db=db,
+        data={
+            "key": fides_key,
+            "name": fides_key,
+            "connection_type": ConnectionType.saas,
+            "access": AccessLevel.write,
+            "secrets": secrets,
+            "saas_config": saas_config,
+        },
+    )
+    yield connection_config
+    connection_config.delete(db)
 
 
 @pytest.fixture(scope="session")
